+++ include/omath/3d_primitives/plane.hpp
@@ -3,8 +3,8 @@
 //
 
 #pragma once
+#include "omath/linear_algebra/vector3.hpp"
 #include "omath/triangle.hpp"
-#include "omath/vector3.hpp"
 #include <array>
 
 namespace omath::primitives
