+++ include/omath/collision/line_tracer.hpp
@@ -3,8 +3,8 @@
 //
 #pragma once
 
+#include "omath/linear_algebra/vector3.hpp"
 #include "omath/triangle.hpp"
-#include "omath/vector3.hpp"
 
 namespace omath::collision
 {
