+++ include/omath/linear_algebra/vector2.hpp
