+++ include/omath/pathfinding/navigation_mesh.hpp
@@ -4,7 +4,7 @@
 
 #pragma once
 
-#include "omath/vector3.hpp"
+#include "omath/linear_algebra/vector3.hpp"
 #include <expected>
 #include <string>
 #include <vector>
