+++ include/omath/projectile_prediction/proj_pred_engine.hpp
@@ -2,9 +2,9 @@
 // Created by Vlad on 2/23/2025.
 //
 #pragma once
+#include "omath/linear_algebra/vector3.hpp"
 #include "omath/projectile_prediction/projectile.hpp"
 #include "omath/projectile_prediction/target.hpp"
-#include "omath/vector3.hpp"
 
 namespace omath::projectile_prediction
 {
