+++ include/omath/projectile_prediction/target.hpp
@@ -3,7 +3,7 @@
 //
 
 #pragma once
-#include "omath/vector3.hpp"
+#include "omath/linear_algebra/vector3.hpp"
 
 namespace omath::projectile_prediction
 {
